+++ cookiecutter/{{cookiecutter.organization}}-richie-site-factory/template/{{cookiecutter.site}}/src/backend/{{cookiecutter.site}}/settings.py
@@ -332,6 +332,25 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
         environ_prefix=None,
     )
 
+    RICHIE_DEFAULT_COURSE_RUN_OFFER = values.Value(
+        "free",
+        environ_name="RICHIE_DEFAULT_COURSE_RUN_OFFER",
+        environ_prefix=None,
+    )
+
+    # Course run price currency value that would be shown on course detail page
+    RICHIE_DEFAULT_COURSE_RUN_PRICE_CURRENCY = values.Value(
+        "EUR",
+        environ_name="RICHIE_DEFAULT_COURSE_RUN_PRICE_CURRENCY",
+        environ_prefix=None,
+    )
+
+    RICHIE_DEFAULT_COURSE_RUN_CERTIFICATE_OFFER = values.Value(
+        "free",
+        environ_name="RICHIE_DEFAULT_COURSE_RUN_CERTIFICATE_OFFER",
+        environ_prefix=None,
+    )
+
     # Internationalization
     TIME_ZONE = "Europe/Paris"
     USE_I18N = True