+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/SyllabusCourseRunCompacted/index.tsx
@@ -41,6 +41,21 @@ const messages = defineMessages({
     description: 'Self paced course run block with no end date',
     defaultMessage: 'Available',
   },
+  coursePrice: {
+    id: 'components.SyllabusCourseRunCompacted.coursePrice',
+    description: 'Title of the course enrollment section of an opened course run block',
+    defaultMessage: 'Enrollment Price',
+  },
+  courseOffer: {
+    id: 'components.SyllabusCourseRunCompacted.courseOffer',
+    description: 'Title of the course enrollment section of an opened course run block',
+    defaultMessage: 'Course offer',
+  },
+  certificationPrice: {
+    id: 'components.SyllabusCourseRunCompacted.certificationPrice',
+    description: 'Title of the course enrollment section of an opened course run block',
+    defaultMessage: 'Certification Price',
+  },
 });
 
 const OpenedSelfPacedCourseRun = ({
@@ -54,6 +69,32 @@ const OpenedSelfPacedCourseRun = ({
   const intl = useIntl();
   const end = courseRun.end ? formatDate(courseRun.end) : '...';
   const hasEndDate = end !== '...';
+  const enrollmentPrice = courseRun?.price != null ? intl.formatNumber(courseRun.price, {
+    style: 'currency',
+    currency: courseRun.price_currency,
+  }) : "...";
+  const certificatePrice = courseRun?.certificate_price != null ? intl.formatNumber(courseRun.certificate_price, {
+    style: 'currency',
+    currency: courseRun.price_currency,
+  }) : "...";
+  const offer = (courseRun.offer ?? "NONE").toUpperCase().replaceAll(" ", "_");
+  const certificationOffer = (courseRun.certificate_offer ?? "NONE").toUpperCase().replaceAll(" ", "_");
+
+  const courseOfferMessage = {
+    "PAID": "The course content is paid.",
+    "FREE": "The course content is free.",
+    "PARTIALLY_FREE": "The course content is free.",
+    "SUBSCRIPTION": "Subscribe to access the course content.",
+    "NONE": ""
+  }[offer]
+
+  const certificationOfferMessage = {
+    "PAID": "The certification process is paid.",
+    "FREE": "The certification process is free.",
+    "SUBSCRIPTION": "The certification process is offered through subscription.",
+    "NONE": ""
+  }[certificationOffer]
+
   return (
     <>
       {courseRun.title && <h3>{StringHelper.capitalizeFirst(courseRun.title)}</h3>}
@@ -81,6 +122,24 @@ const OpenedSelfPacedCourseRun = ({
               <FormattedMessage {...messages.languages} />
             </dt>
             <dd>{IntlHelper.getLocalizedLanguages(courseRun.languages, intl)}</dd>
+          </>
+        )}
+        {courseRun?.price != null && (
+          <>
+            <dt>
+              <FormattedMessage {...messages.coursePrice} />
+            </dt>
+            <dd>
+              {`${courseOfferMessage}`}<br/>{`${enrollmentPrice}`}
+            </dd>
+          </>
+        )}
+        { courseRun?.certificate_price != null && (
+          <>
+            <dt>
+              <FormattedMessage {...messages.certificationPrice} />
+            </dt>
+            <dd>{`${certificationOfferMessage}`}<br/>{`${certificatePrice}`}</dd>
           </>
         )}
       </dl>