+++ src/richie/apps/courses/admin.py
@@ -57,6 +57,8 @@ class CourseRunAdminForm(TranslatableModelForm):
             "languages",
             "enrollment_count",
             "catalog_visibility",
+            "offer",
+            "price",
             "sync_mode",
             "display_mode",
         ]
@@ -150,6 +152,8 @@ class CourseRunAdmin(FrontendEditableAdminMixin, TranslatableAdmin):
         "languages",
         "enrollment_count",
         "catalog_visibility",
+        "offer",
+        "price",
         "sync_mode",
     )
     list_display = ["id"]